examples/text_to_agent/web_search/gemini_live/pyproject.toml
[project]
name = "web-search-gemini-live"
version = "0.1.0"
description = "Web search agent using Gemini Live"
dependencies = [
    "cartesia-line",
    "google-genai",
    "loguru",
    "fastapi",
    "uvicorn",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
